% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grch37}
\alias{grch37}
\title{GRCh37 chromosome lengths}
\format{
A named numeric vector of length 25.
}
\source{
\url{http://hgdownload.soe.ucsc.edu/goldenPath/hg19/bigZips/hg19.chrom.sizes}
}
\usage{
grch37
}
\description{
A named numeric vector of chromosome lengths for the GRCh37 (hg19)
human genome assembly. Names are "1"-"22", "X", "Y", "MT".
}
\keyword{datasets}
