% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grch38}
\alias{grch38}
\title{GRCh38 chromosome lengths}
\format{
A named numeric vector of length 25.
}
\source{
\url{http://hgdownload.soe.ucsc.edu/goldenPath/hg38/bigZips/hg38.chrom.sizes}
}
\usage{
grch38
}
\description{
A named numeric vector of chromosome lengths for the GRCh38 (hg38)
human genome assembly. Names are "1"-"22", "X", "Y", "MT".
}
\keyword{datasets}
