% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_serial.R
\name{sort_domains}
\alias{sort_domains}
\title{Sort domains in natural order}
\usage{
sort_domains(domains)
}
\arguments{
\item{domains}{Character vector of domain names.}
}
\value{
Character vector of unique domains in sorted order.
}
\description{
Numeric domains first, then known genomic domains (X, Y, MT), then
alphabetical.
}
\keyword{internal}
